# packages/api-client/openapi/servier-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory/{ingredientId}": {
      "put": {
        "operationId": "PotionLabController_setInventoryQuantity",
        "parameters": [
          {
            "name": "ingredientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantity"
                ],
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated inventory item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid quantity."
          },
          "404": {
            "description": "Unknown ingredient."
          }
        },
        "summary": "Set an ingredient inventory quantity.",
        "tags": [
          "Potion Lab"
        ]
      }
    },
    "/inventory/{ingredientId}/recharge": {
      "post": {
        "operationId": "PotionLabController_recharge",
        "parameters": [
          {
            "name": "ingredientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The recharged inventory item.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid recharge amount."
          },
          "404": {
            "description": "Unknown ingredient."
          }
        },
        "summary": "Increase one ingredient quantity.",
        "tags": [
          "Potion Lab"
        ]
      }
    },
    "/inventory/randomize": {
      "post": {
        "operationId": "PotionLabController_randomizeInventory",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "minimum": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 1
                  },
                  "maximum": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The randomized inventory.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItemDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid randomization bounds."
          }
        },
        "summary": "Randomize inventory quantities for game replay.",
        "tags": [
          "Potion Lab"
        ]
      }
    },
...
    "/potions": {
      "get": {
        "operationId": "PotionLabController_listPotions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Potion creation history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PotionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List created potion history.",
        "tags": [
          "Potion Lab"
        ]
      },
      "post": {
        "operationId": "PotionLabController_createPotion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ingredientIds"
                ],
                "properties": {
                  "ingredientIds": {
                    "type": "array",
                    "minItems": 3,
                    "maxItems": 3,
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created potion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotionDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid selection, missing recipe, or insufficient stock."
          }
        },
        "summary": "Create a potion from exactly three ingredient IDs.",
        "tags": [
          "Potion Lab"
        ]
      }
    }
  },
  "info": {
    "title": "SERVIER Potion Lab API",
    "description": "Inventory, recipe, and potion endpoints for the SERVIER technical test.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "schemas": {
...
      "InventoryItemDto": {
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": "string",
            "example": "noix-de-coco"
          },
          "name": {
            "type": "string",
            "example": "Noix de coco"
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "example": 3
          }
        },
        "required": [
          "ingredientId",
          "name",
          "quantity"
        ]
      },
...
      "PotionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "4a8d4d42-2c47-4a3a-9e28-9639dc7c5f52"
          },
          "recipeId": {
            "type": "string",
            "example": "invisibilite"
          },
          "name": {
            "type": "string",
            "example": "Potion d'invisibilité"
          },
          "ingredientIds": {
            "minItems": 3,
            "maxItems": 3,
            "example": [
              "noix-de-coco",
              "yttrium",
              "mandragore"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-21T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "recipeId",
          "name",
          "ingredientIds",
          "createdAt"
        ]
      }
    }
  }
}